+++ inc/ariajanke/cul/Either.hpp
@@ -27,89 +27,61 @@
 #pragma once
 
 #include <ariajanke/cul/OptionalEither.hpp>
+#include <ariajanke/cul/Util.hpp>
+#include <ariajanke/cul/EitherFold.hpp>
 
 namespace cul {
 
 template <typename LeftT, typename RightT>
-class Either final : public OptionalEitherBase {
+class Either final : public detail::EitherBase {
 public:
     using LeftType = LeftT;
     using RightType = RightT;
 
 private:
     template <typename EitherLeftOrRight>
-    using EnableIfLeftXorRightPtr = typename std::enable_if_t<
-        bool(std::is_same_v<EitherLeftOrRight, LeftType > ^
-             std::is_same_v<EitherLeftOrRight, RightType>),
-        void *>;
+    using EnableIfLeftXorRightPtr =
+        EnableIfLeftXorRightPtr_<LeftType, RightType, EitherLeftOrRight>;
 
 public:
-
-    constexpr explicit Either(OptionalEither<LeftT, RightT> && opt_either):
-        m_datum(verify_non_empty("Either", std::move(opt_either))) {}
+    constexpr explicit Either(OptionalEither<LeftT, RightT> && opt_either);
 
     template <typename EitherLeftOrRight>
-    constexpr Either(EitherLeftOrRight && obj,
-                     EnableIfLeftXorRightPtr<EitherLeftOrRight> = nullptr):
-        m_datum(std::move(obj)) {}
+    constexpr explicit Either(EitherLeftOrRight && obj,
+                              EnableIfLeftXorRightPtr<EitherLeftOrRight> = nullptr);
 
-    constexpr Either(TypeTag<LeftType>, RightType && right):
-        m_datum(TypeTag<LeftType>{}, std::move(right)) {}
+    constexpr Either(TypeTag<LeftType>, RightType && right);
 
-    constexpr Either(LeftType && left, TypeTag<RightType>):
-        m_datum(std::move(left), TypeTag<RightType>{}) {}
+    constexpr Either(LeftType && left, TypeTag<RightType>);
 
     template <typename CommonT>
-    constexpr Fold<LeftType, RightType, CommonT> fold()
-        { return FoldAttn::make_fold<LeftType, RightType, CommonT>(std::move(m_datum)); }
+    constexpr either::Fold<LeftType, RightType, CommonT> fold();
 
     template <typename Func>
-    [[nodiscard]] constexpr Either<LeftType, ReturnTypeOf<Func>> map(Func && f)
-        { return Either<LeftType, ReturnTypeOf<Func>>{m_datum.template map<Func>(std::move(f))}; }
+    [[nodiscard]] constexpr Either<LeftType, ReturnTypeOf<Func>>
+        map(Func && f);
 
     template <typename Func>
-    [[nodiscard]] constexpr Either<ReturnTypeOf<Func>, RightType> map_left(Func && f)
-        { return Either<ReturnTypeOf<Func>, RightType>{m_datum.template map_left<Func>(std::move(f))}; }
+    [[nodiscard]] constexpr Either<ReturnTypeOf<Func>, RightType>
+        map_left(Func && f);
 
     template <typename Func>
-    [[nodiscard]] constexpr EnableIfReturnsEither<Func> chain(Func && f) {
-        RightChainFunctionReturnRequirements<Func>{};
-        VerifyArgumentsForSide<Func, RightType>{};
-        if (is_right())
-            { return f(right()); }
-        return with_new_right_type<EitherTypeRight<ReturnTypeOf<Func>>>();
-    }
+    [[nodiscard]] constexpr EnableIfReturnsEither<Func> chain(Func && f);
 
     template <typename Func>
-    [[nodiscard]] constexpr EnableIfReturnsEither<Func> chain_left(Func && f) {
-        LeftChainFunctionReturnRequirements<Func>{};
-        VerifyArgumentsForSide<Func, LeftType>{};
-        if (is_left())
-            { return f(left()); }
-        return with_new_left_type<EitherTypeLeft<ReturnTypeOf<Func>>>();
-    }
+    [[nodiscard]] constexpr EnableIfReturnsEither<Func> chain_left(Func && f);
 
-    constexpr LeftT left()
-        { return m_datum.left(); }
+    constexpr LeftT left();
 
-    constexpr RightT right()
-        { return m_datum.right(); }
+    constexpr RightT right();
 
-    constexpr bool is_left() const
-        { return m_datum.is_left(); }
+    constexpr bool is_left() const;
 
-    constexpr bool is_right() const
-        { return m_datum.is_right(); }
+    constexpr bool is_right() const;
 
 private:
     static constexpr OptionalEither<LeftT, RightT> verify_non_empty
-        (const char * caller, OptionalEither<LeftT, RightT> && opt_either)
-    {
-        using namespace exceptions_abbr;
-        if (!opt_either.is_empty()) return opt_either;
-        throw RtError{"Either::" + std::string{caller} +
-                      ": given optional either must not be empty."};
-    }
+        (const char * caller, OptionalEither<LeftT, RightT> && opt_either);
 
     template <typename Func>
     struct RightChainFunctionReturnRequirements final :
@@ -124,18 +96,10 @@ private:
     {};
 
     template <typename NewRightType>
-    Either<LeftType, NewRightType> with_new_right_type() {
-        return Either<LeftType, NewRightType>
-            {OptionalEitherToEitherAttn::new_right_type_with
-                <LeftType, RightType, NewRightType>(std::move(m_datum))};
-    }
+    Either<LeftType, NewRightType> with_new_right_type();
 
     template <typename NewLeftType>
-    Either<NewLeftType, RightType> with_new_left_type() {
-        return Either<NewLeftType, RightType>
-            {OptionalEitherToEitherAttn::new_left_type_with
-                <LeftType, RightType, NewLeftType>(std::move(m_datum))};
-    }
+    Either<NewLeftType, RightType> with_new_left_type();
 
     OptionalEither<LeftT, RightT> m_datum;
 };
@@ -179,5 +143,135 @@ constexpr EitherLeftMaker<LeftType> left(LeftType && obj) {
 }
 
 } // end of either namespace -> into ::cul
+
+// ----------------------------------------------------------------------------
+
+template <typename LeftT, typename RightT>
+constexpr Either<LeftT, RightT>::
+    Either(OptionalEither<LeftT, RightT> && opt_either):
+    m_datum(verify_non_empty("Either", std::move(opt_either))) {}
+
+template <typename LeftT, typename RightT>
+template <typename EitherLeftOrRight>
+constexpr Either<LeftT, RightT>::Either
+    (EitherLeftOrRight && obj,
+     EnableIfLeftXorRightPtr<EitherLeftOrRight>):
+    m_datum(std::move(obj)) {}
+
+template <typename LeftT, typename RightT>
+constexpr Either<LeftT, RightT>::Either(TypeTag<LeftType>, RightType && right):
+    m_datum(TypeTag<LeftType>{}, std::move(right)) {}
+
+template <typename LeftT, typename RightT>
+constexpr Either<LeftT, RightT>::Either(LeftType && left, TypeTag<RightType>):
+    m_datum(std::move(left), TypeTag<RightType>{}) {}
+
+template <typename LeftT, typename RightT>
+template <typename CommonT>
+constexpr either::Fold<LeftT, RightT, CommonT>
+    Either<LeftT, RightT>::fold()
+{
+    return detail::FoldAttn::make_fold<LeftType, RightType, CommonT>
+        (std::move(m_datum));
+}
+
+template <typename LeftT, typename RightT>
+template <typename Func>
+[[nodiscard]] constexpr
+    Either<LeftT,
+           typename Either<LeftT, RightT>::
+               template ReturnTypeOf<Func>>
+    Either<LeftT, RightT>::map(Func && f)
+{
+    return Either<LeftType, ReturnTypeOf<Func>>
+        {m_datum.template map<Func>(std::move(f))};
+}
+
+template <typename LeftT, typename RightT>
+template <typename Func>
+[[nodiscard]] constexpr
+    Either<typename Either<LeftT, RightT>::
+               template ReturnTypeOf<Func>,
+           RightT>
+    Either<LeftT, RightT>::map_left(Func && f)
+{
+    return Either<ReturnTypeOf<Func>, RightType>
+        {m_datum.template map_left<Func>(std::move(f))};
+}
+
+template <typename LeftT, typename RightT>
+template <typename Func>
+[[nodiscard]] constexpr
+    typename Either<LeftT, RightT>::template EnableIfReturnsEither<Func>
+    Either<LeftT, RightT>::chain(Func && f)
+{
+    RightChainFunctionReturnRequirements<Func>{};
+    VerifyArgumentsForSide<Func, RightType>{};
+    if (is_right())
+        { return f(right()); }
+    return with_new_right_type<EitherTypeRight<ReturnTypeOf<Func>>>();
+}
+
+template <typename LeftT, typename RightT>
+template <typename Func>
+[[nodiscard]] constexpr
+    typename Either<LeftT, RightT>::template EnableIfReturnsEither<Func>
+    Either<LeftT, RightT>::chain_left(Func && f)
+{
+    LeftChainFunctionReturnRequirements<Func>{};
+    VerifyArgumentsForSide<Func, LeftType>{};
+    if (is_left())
+        { return f(left()); }
+    return with_new_left_type<EitherTypeLeft<ReturnTypeOf<Func>>>();
+}
+
+template <typename LeftT, typename RightT>
+constexpr LeftT Either<LeftT, RightT>::left()
+    { return m_datum.left(); }
+
+template <typename LeftT, typename RightT>
+constexpr RightT Either<LeftT, RightT>::right()
+    { return m_datum.right(); }
+
+template <typename LeftT, typename RightT>
+constexpr bool Either<LeftT, RightT>::is_left() const
+    { return m_datum.is_left(); }
+
+template <typename LeftT, typename RightT>
+constexpr bool Either<LeftT, RightT>::is_right() const
+    { return m_datum.is_right(); }
+
+template <typename LeftT, typename RightT>
+/* private static */ constexpr OptionalEither<LeftT, RightT>
+    Either<LeftT, RightT>::verify_non_empty
+    (const char * caller, OptionalEither<LeftT, RightT> && opt_either)
+{
+    using namespace exceptions_abbr;
+    if (!opt_either.is_empty()) return opt_either;
+    throw RtError{"Either::" + std::string{caller} +
+                  ": given optional either must not be empty."};
+}
+
+template <typename LeftT, typename RightT>
+template <typename NewRightType>
+/* private */ Either<LeftT, NewRightType>
+    Either<LeftT, RightT>::with_new_right_type()
+{
+    using namespace detail;
+    return Either<LeftType, NewRightType>
+        {OptionalEitherToEitherAttn::new_right_type_with
+            <LeftType, RightType, NewRightType>(std::move(m_datum))};
+}
+
+template <typename LeftT, typename RightT>
+template <typename NewLeftType>
+/* private */ Either<NewLeftType, RightT>
+    Either<LeftT, RightT>::with_new_left_type()
+{
+    using namespace detail;
+    return Either<NewLeftType, RightType>
+        {OptionalEitherToEitherAttn::new_left_type_with
+            <LeftType, RightType, NewLeftType>(std::move(m_datum))};
+}
 
 } // end of cul namespace
